requirements.txt
requests==2.28.1
marshmallow==3.19.0
titlecase==2.4

# https://github.com/bear/python-twitter
python-twitter==3.5
